requirements.txt
numpy==2.2.5
PyQt5==5.15.11
pyautogui==0.9.54
keyboard==0.13.5
PyYAML==6.0.2
opencv-python~=4.12.0.88
pylint==3.3.4
black==25.1.0
isort==6.0.1
pre-commit==4.3.0